refactor, flatten, add TODOs
