+++ src/main/java/org/null0/edi/debatcher/Debatcher.java
@@ -17,12 +17,9 @@ import java.util.Stack;
 import org.apache.commons.lang3.StringUtils;
 import org.null0.edi.debatcher.DebatcherException.ERROR_LEVEL;
 import org.null0.edi.debatcher.DebatcherException.ERROR_OR_EXCEPTION;
-import org.null0.edi.debatcher.model.HierarchicalLevel;
-import org.null0.edi.debatcher.validation.EncounterEdiValidator;
-import org.null0.edi.debatcher.validation.EncounterEdiValidatorImpl;
-import org.null0.edi.debatcher.validation.EncounterEdiValidator.CLAIM_TYPE;
-import org.null0.edi.debatcher.validation.EncounterEdiValidator.ERROR;
-import org.null0.edi.debatcher.validation.EncounterEdiValidator.X12_837_ELEMENT;
+import org.null0.edi.debatcher.EncounterEdiValidator.CLAIM_TYPE;
+import org.null0.edi.debatcher.EncounterEdiValidator.ERROR;
+import org.null0.edi.debatcher.EncounterEdiValidator.X12_837_ELEMENT;
 import org.null0.edi.debatcher.MetadataLogger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
